pyproject.toml
[project]
name = "research-rabbit"
version = "0.0.1"
description = "Lightweight web research and summarization assistant."
authors = [
    { name = "Lance Martin" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
    "langgraph>=0.2.55",             # Required for StateGraph
    "langchain-community>=0.3.9",
    "langchain-ollama>=0.2.1",       # Integration with ChatOllama
    "langchain-pinecone>=0.0.2",     # Pinecone integration with LangChain
    "langchain-openai>=0.0.5",       # OpenAI embeddings integration
    "pinecone-client>=2.2.4",        # Pinecone client for vector storage
    "protobuf>=4.21.0",              # Required for Pinecone compatibility
    "tavily-python>=0.5.0",          # For Tavily search API
    "typing-extensions>=4.10.0",     # Ensures compatibility with Literal typing
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["research_rabbit"]

[tool.setuptools.package-dir]
"research_rabbit" = "src/research_rabbit"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"